refactor(workflows): rename /single-story-plan to /story-plan (refs #3705)
Hard cutover for command-surface symmetry with /story-deliver and the
/epic-plan -> /epic-deliver pairing. Renames the workflow, the CLI script
and its lib helper, the test file, and the context envelope kind
(single-story-plan-context -> story-plan-context). Repoints SDLC.md and
both READMEs, and registers single-story-plan in check-doc-links.js's
RETIRED_COMMANDS so the old name cannot reappear in active docs. The
.claude/commands projection is regenerated at runtime via sync:commands
(gitignored).
